+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### 0.4.8 (2018.12.27)
 * Reorganize python files into directories:
-    * path updated with: `import update_sys_path`
+    * path updated with: import update_sys_path
     * columns.py and custom_titles.py now have code wrapped in a class
     * Multiple simultaneous sessions enabled again by wrapping all bokeh objects into classes
 * Remove test files
@@ -14,6 +14,7 @@
 * All sql/preference backups stored in ~/Apps/dvh_analytics/data/backup now (can't customize)
 * options.py now contains imports (os and paths.py)
     * This broke load_options, code added to ignore ModuleType
+* Automatically update uncategorized variations in ROI Manager after importing data
 
 ### 0.4.7 (2018.12.6)
 * Move csv creation to python for less javascript (download.js)